requirements.txt
requests==2.32.3
python-dotenv==1.0.1
langflow==1.1.4.post1
streamlit==1.42.0